requirements.txt
ccxt==4.2.85
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.2.1
ta==0.11.0